build-system/haskell: Simplify configure step.
The "--extra-include-dirs" and "--extra-lib-dirs" options end up adding
references to too many unnecessary inputs.  Turns out that these options
aren't even needed for a correct build.

The "--bindir" option is not as useful as it seems as the configured location
is embedded in the outputs.  Instead of using "--bindir" it seems better to
build a statically linked binary and move the binary to its own output to
avoid references between the "out" and "bin" outputs.

* guix/build/haskell-build-system.scm (configure): Do not pass "--bindir",
"--extra-include-dirs", and "--extra-lib-dirs".