import: git: Restrict to non-github origins.
'guix refresh -t generic-git' reported a lot of packages which are hosted at
github and are already handled by the github updater.  So whether the package
is updates using the github updater or the generic-git updater would be
non-deterministic.

And indeed 'guix refresh --list-updaters' without this change reported 31.2%
for the github updater and 34.9% for the generic-git updater.  With this
patch, it's only 3.7% for the generic-git updater.  (Still the sum of the
percent-values is 112%, anyhow better then the 146% prior to this change.)

Beside these statistical issues, this patch has another effect:  If the github
updater fails to pick a (later) version the generic-git updater was
triggered.  Anyhow: Why should the generic-git updater be able to find
a (later) version, if the github updater did not?

* guix/import/github.scm: Export 'github-package?'.
* guix/import/git.scm(git-package?): Also check for not github-package?.